understand/data-lineage/marquez-demo/requirements.txt
attrs==23.2.0
aws-cdk-lib==2.142.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
aws-cdk.lambda-layer-kubectl-v29==2.0.0
cattrs==23.2.3
constructs==10.3.0
exceptiongroup==1.2.1
importlib_resources==6.4.0
jsii==1.98.0
publication==0.0.3
python-dateutil==2.9.0.post0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.11.0
zipp==3.18.1
